+++ src/main/java/eu/de4a/connector/mock/exampledata/ExampleDataException.java
@@ -1,0 +1,8 @@
+package eu.de4a.connector.mock.exampledata;
+
+public class ExampleDataException extends IllegalArgumentException {
+
+    ExampleDataException(String message) {
+        super(message);
+    }
+}